requirements.txt
asgiref==3.7.2
astroid==3.0.0
colorama==0.4.6
dill==0.3.7
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
Faker==19.6.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.11.0
pylint==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.12.1
typed-ast==1.5.5
tzdata==2023.3
validate-docbr==1.10.0
wrapt==1.15.0
